+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Change Log
 
 ## [Unreleased] - 2017-02-21
+- Add `Chars.IsHexDigit()` function.
 - First release with Changelog.